move user data models
